Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main(){
-    int average;
+    float average;
     int a,b,c;
     printf("");
     scanf("%d %d %d",&a,&b,&c);
     average = (a+b+c)/3;
-    printf("Average: %d\n",average);
+    printf("Average: %.2f\n",average);
     return 0;
 }
